+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- logical builtin added:
+  `logical`, `true`,
+  `false`, `"xor`,
+
 - single builtin added:
   `single`